fix bug

+++ include/memmap.h
@@ -9,8 +9,8 @@
  *  qemu does not support raspi4
  */
 #define PERIPHERAL_BASE 0xfe000000
-//#define PERIPHERAL_BASE 0x3f000000
+// #define PERIPHERAL_BASE 0x3f000000
 
-#define REG(addr) (*(volatile u32 *)(addr))
+#define REG(addr) (*(volatile u32 *)(u64)(addr))
 
 #endif
